pyproject.toml
[project]
name = "kcmh-sql-bot"
version = "0.1.0"
description = "Read-only analytics chatbot for KCMH Hospital Information System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "anthropic>=0.40.0",
    "psycopg[binary]>=3.2.0",
    "sqlglot>=26.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg-pool>=3.3.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
